requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotli-python=1.0.9=py38hfa26641_7
btf-extractor=1.7.0=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py38h06a4308_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.37=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h0708190_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
h5py=3.11.0=pypi_0
icu=67.1=he1b5a44_0
idna=3.7=pyhd8ed1ab_0
imageio=2.34.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h166bdaf_1
kiwisolver=1.4.4=py38h6a678d5_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=hf544144_3
libwebp-base=1.2.2=h7f98852_1
lz4-c=1.9.3=h9c3ff4c_1
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h5d868c9_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
ncurses=6.4=h6a678d5_0
nettle=3.6=he412f7d_0
nptyping=1.4.4=pypi_0
numpy=1.24.3=py38h14f4228_0
numpy-base=1.24.3=py38h31eccc5_0
olefile=0.47=pyhd8ed1ab_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openexr=3.2.4=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1w=h7f8727e_0
pillow=10.3.0=pypi_0
pip=24.0=py38h06a4308_0
prefetch-generator=1.0.3=pypi_0
pynvml=11.5.0=pypi_0
pyparsing=3.0.9=py38h06a4308_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.1=h357f687_2
python-dateutil=2.9.0post0=py38h06a4308_0
python_abi=3.8=2_cp38
pytorch=1.12.1=py3.8_cuda11.6_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
readline=8.2=h5eee18b_0
requests=2.31.0=pyhd8ed1ab_0
setuptools=69.5.1=py38h06a4308_0
simplejpeg=1.7.2=pypi_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h5eee18b_0
tk=8.6.14=h39e8969_0
torchaudio=0.12.1=py38_cu116
torchvision=0.13.1=py38_cu116
tornado=6.3.3=py38h5eee18b_0
tqdm=4.66.4=pypi_0
typing_extensions=4.11.0=pyha770c72_0
typish=1.9.3=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=pyhd8ed1ab_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.0=ha95c52a_0
